General refactoring
Move classes to different namespaces
