+++ Transmission.API.RPC/Api/Entity/RenameTorrentInfo.cs
@@ -1,0 +1,27 @@
+﻿using Newtonsoft.Json;
+
+namespace Transmission.API.RPC.Api.Entity;
+
+/// <summary>
+/// Rename torrent result information
+/// </summary>
+public class RenameTorrentInfo
+{
+    /// <summary>
+    /// The torrent's unique Id.
+    /// </summary>
+    [JsonProperty("id")]
+    public int ID { get; set; }
+
+    /// <summary>
+    /// File path.
+    /// </summary>
+    [JsonProperty("path")]
+    public string Path { get; set; }
+
+    /// <summary>
+    /// File name.
+    /// </summary>
+    [JsonProperty("name")]
+    public string Name { get; set; }
+}
