+++ Transmission.API.RPC/Api/Entity/Torrent/TorrentFileStats.cs
@@ -1,0 +1,11 @@
+﻿using Transmission.API.RPC.Core.Entity.Torrent;
+using Transmission.API.RPC.Core.Enums;
+
+namespace Transmission.API.RPC.Api.Entity.Torrent;
+
+public class TorrentFileStats : ITorrentFileStats
+{
+    public double BytesCompleted { get; set; }
+    public bool Wanted { get; set; }
+    public Priority Priority { get; set; }
+}
