+++ Transmission.API.RPC/Api/TransmissionResponse.cs
@@ -1,0 +1,21 @@
+﻿using Newtonsoft.Json;
+
+namespace Transmission.API.RPC.Api;
+
+/// <summary>
+/// Transmission response 
+/// </summary>
+public class TransmissionResponse : CommunicateBase
+{
+    [JsonConstructor]
+    internal TransmissionResponse(string result)
+    {
+        Result = result;
+    }
+
+    /// <summary>
+    /// Contains "success" on success, or an error string on failure.
+    /// </summary>
+    [JsonProperty("result")]
+    public string Result;
+}
