--- Transmission.API.RPC/Common/Enums.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Transmission.API.RPC.Common
-{
-    public enum Priority
-    {
-        Low = -1,
-        Normal = 0,
-        High = 1
-    }
-}
