+++ Transmission.API.RPC/Core/Enums/Priority.cs
@@ -1,0 +1,8 @@
+﻿namespace Transmission.API.RPC.Core.Enums;
+
+public enum Priority
+{
+    Low = -1,
+    Normal = 0,
+    High = 1
+}
