+++ Transmission.API.RPC/Core/Util/AsyncExtensions.cs
@@ -1,0 +1,28 @@
+﻿namespace Transmission.API.RPC.Core.Util;
+
+/// <summary>
+/// Async extension
+/// </summary>
+public static class AsyncExtensions
+{
+    /// <summary>
+    /// Wait and unwrap exception
+    /// </summary>
+    /// <param name="task"></param>
+    public static void WaitAndUnwrapException(this Task task)
+    {
+        try
+        {
+            task.Wait();
+        }
+        catch (Exception e)
+        {
+            if (e.InnerException != null)
+            {
+                throw e.InnerException;
+            }
+
+            throw;
+        }
+    }
+}
